requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
lxml==5.1.0
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
openpyxl==3.1.2
pandas==2.1.4
eventlet==0.35.1
